Hackish solution to deal with missing info
The current database is missing information about voting
locations.  This is available from a separate source (PDF!).

Take the PDF file as input, massage it to produce tabular data
and import that data into the database.

Once the information becomes available from the database, this
ought to be refactored so that it's used only in special
circumstances.